Ready for PR on #100. Good enough I think.
